:twisted_rightwards_arrows: funcional Navbar and Sidebar
